fixing race conditions
